moving CodeCommand to /command
