+++ src/main/java/com/github/hokkaydo/eplbot/module/code/CodeModule.java
@@ -2,6 +2,7 @@ package com.github.hokkaydo.eplbot.module.code;
 
 import com.github.hokkaydo.eplbot.command.Command;
 import com.github.hokkaydo.eplbot.module.Module;
+import com.github.hokkaydo.eplbot.module.code.command.CodeCommand;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.jetbrains.annotations.NotNull;
